debug some comments
