+++ src/main.js
@@ -1,5 +1,7 @@
 'use strict';
-import calculateExpression from './calculateExpression.js';
+
+// Виписати окреми функції логіки
+// document.addEventListener('keyup', removeClasslist)
 
 const displayMain = document.getElementById('displayMain');
 
@@ -132,6 +134,66 @@ document.addEventListener('keyup', event => {
         return;
 
       // Working with results
+
+      function calculateExpression(expression) {
+        // Replacing '×' and '÷' to * and /
+
+        expression = expression.replace(/×/g, '*').replace(/÷/g, '/');
+
+        // Parse string in the comfortable array format to work with
+
+        function splitExpressionString(expression) {
+          let counter = 0;
+          let separatedString = '';
+          const operators = ['+', '-', '*', '/'];
+
+          if (expression[0].startsWith('-')) {
+            counter = 1;
+            // Parsing anyway the first operand if it is negative (starts with the '-' operator)
+            separatedString += expression[0];
+          }
+          for (counter; counter < expression.length; counter++) {
+            !operators.includes(expression[counter]) ? (separatedString += expression[counter]) : (separatedString += ` ${expression[counter]} `);
+          }
+          return separatedString;
+        }
+
+        const parsedExpression = splitExpressionString(expression);
+
+        const expressionArray = parsedExpression.split(' ');
+        // mutating expression array with calculating results * and / operands pairs
+
+        function multiplyOrDevideOperands(expression, operator) {
+          while (expression.includes(`${operator}`)) {
+            for (let i = 1; i < expression.length; i += 2) {
+              const currentOperator = expression[i];
+              if (currentOperator === `${operator}`) {
+                if (operator === '*') {
+                  const multiplyResult = expression[i - 1] * expression[i + 1];
+                  expression.splice(i - 1, 3, multiplyResult);
+                } else if (operator === '/') {
+                  const devideResult = expression[i - 1] / expression[i + 1];
+                  expression.splice(i - 1, 3, devideResult);
+                }
+              }
+            }
+          }
+        }
+
+        multiplyOrDevideOperands(expressionArray, '*');
+        multiplyOrDevideOperands(expressionArray, '/');
+
+        // calculating result
+
+        let result = parseFloat(expressionArray[0]);
+
+        for (let i = 1; i < expressionArray.length; i += 2) {
+          const operator = expressionArray[i];
+          const operand = parseFloat(expressionArray[i + 1]);
+          operator === '+' ? (result += operand) : (result -= operand);
+        }
+        return result;
+      }
 
       const expressionResult = calculateExpression(displayValue);
